# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grateful – Gratitude Journal API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

/**
 * @nullable
 */
export type EntryMood = typeof EntryMood[keyof typeof EntryMood] | null;


export const EntryMood = {
  great: 'great',
  okay: 'okay',
  tough: 'tough',
} as const;

export interface Entry {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** 1-3 gratitude items */
  gratitudeItems: string[];
  reflection: string;
  /** @nullable */
  mood: EntryMood;
  starred: boolean;
  categories: string[];
  /** @nullable */
  photoPath?: string | null;
  createdAt: string;
}

/**
 * @nullable
 */
export type EntryInputMood = typeof EntryInputMood[keyof typeof EntryInputMood] | null;


export const EntryInputMood = {
  great: 'great',
  okay: 'okay',
  tough: 'tough',
} as const;

export interface EntryInput {
  date: string;
  /**
     * @minItems 1
     * @maxItems 3
     */
  gratitudeItems: string[];
  reflection: string;
  /** @nullable */
  mood?: EntryInputMood;
  categories?: string[];
}

/**
 * @nullable
 */
export type EntryUpdateMood = typeof EntryUpdateMood[keyof typeof EntryUpdateMood] | null;


export const EntryUpdateMood = {
  great: 'great',
  okay: 'okay',
  tough: 'tough',
} as const;

export interface EntryUpdate {
  /**
     * @minItems 1
     * @maxItems 3
     */
  gratitudeItems?: string[];
  reflection?: string;
  /** @nullable */
  mood?: EntryUpdateMood;
  starred?: boolean;
  categories?: string[];
}

export type WeeklySummaryMoodBreakdown = {
  great: number;
  okay: number;
  tough: number;
  untagged: number;
};

export type WeeklySummaryCategoryBreakdown = {[key: string]: number};

export interface WeeklySummary {
  startDate: string;
  endDate: string;
  entries: Entry[];
  totalEntries: number;
  moodBreakdown: WeeklySummaryMoodBreakdown;
  allGratitudeItems: string[];
  categoryBreakdown: WeeklySummaryCategoryBreakdown;
}

export interface OnThisDay {
  oneWeekAgo: Entry | null;
  oneMonthAgo: Entry | null;
  oneYearAgo: Entry | null;
}

export interface Streak {
  currentStreak: number;
  longestStreak: number;
  todayLogged: boolean;
}

export interface DailyPrompt {
  prompt: string;
}

export interface ErrorResponse {
  error: string;
}