# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Grateful – Gratitude Journal API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: entries
    description: Journal entry operations
  - name: prompts
    description: Writing prompt operations
  - name: streak
    description: Streak operations
  - name: summary
    description: Weekly summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      security:
        - bearerAuth: []
      summary: List all journal entries
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
    post:
      operationId: createEntry
      tags: [entries]
      security:
        - bearerAuth: []
      summary: Create a new journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryInput"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Entry for this date already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /entries/favorites:
    get:
      operationId: getFavorites
      tags: [entries]
      security:
        - bearerAuth: []
      summary: Get all starred/favorite entries
      responses:
        "200":
          description: Favorite entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"

  /entries/onthisday:
    get:
      operationId: getOnThisDay
      tags: [entries]
      security:
        - bearerAuth: []
      summary: Get entries from 1 week, 1 month, and 1 year ago
      responses:
        "200":
          description: On this day entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnThisDay"

  /entries/week/{startDate}:
    get:
      operationId: getWeeklySummary
      tags: [summary]
      security:
        - bearerAuth: []
      summary: Get weekly summary starting from a given date
      parameters:
        - name: startDate
          in: path
          required: true
          schema:
            type: string
            description: Start date in YYYY-MM-DD format (typically Sunday)
      responses:
        "200":
          description: Weekly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklySummary"

  /entries/{date}:
    get:
      operationId: getEntryByDate
      tags: [entries]
      security:
        - bearerAuth: []
      summary: Get journal entry by date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEntry
      tags: [entries]
      security:
        - bearerAuth: []
      summary: Update an existing journal entry
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryUpdate"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /streak:
    get:
      operationId: getStreak
      tags: [streak]
      security:
        - bearerAuth: []
      summary: Get current journaling streak
      responses:
        "200":
          description: Streak data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Streak"

  /prompts/daily:
    get:
      operationId: getDailyPrompt
      tags: [prompts]
      summary: Get today's writing prompt
      responses:
        "200":
          description: Daily prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPrompt"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    AuthInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string

    Entry:
      type: object
      required: [id, date, gratitudeItems, reflection, mood, starred, categories, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
          description: Date in YYYY-MM-DD format
        gratitudeItems:
          type: array
          items:
            type: string
          description: 1-3 gratitude items
        reflection:
          type: string
        mood:
          type: ["string", "null"]
          enum: ["great", "okay", "tough", null]
        starred:
          type: boolean
        categories:
          type: array
          items:
            type: string
        photoPath:
          type: ["string", "null"]
        createdAt:
          type: string

    EntryInput:
      type: object
      required: [date, gratitudeItems, reflection]
      properties:
        date:
          type: string
        gratitudeItems:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 3
        reflection:
          type: string
        mood:
          type: ["string", "null"]
          enum: ["great", "okay", "tough", null]
        categories:
          type: array
          items:
            type: string

    EntryUpdate:
      type: object
      properties:
        gratitudeItems:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 3
        reflection:
          type: string
        mood:
          type: ["string", "null"]
          enum: ["great", "okay", "tough", null]
        starred:
          type: boolean
        categories:
          type: array
          items:
            type: string

    WeeklySummary:
      type: object
      required: [startDate, endDate, entries, totalEntries, moodBreakdown, allGratitudeItems, categoryBreakdown]
      properties:
        startDate:
          type: string
        endDate:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
        totalEntries:
          type: integer
        moodBreakdown:
          type: object
          required: [great, okay, tough, untagged]
          properties:
            great:
              type: integer
            okay:
              type: integer
            tough:
              type: integer
            untagged:
              type: integer
        allGratitudeItems:
          type: array
          items:
            type: string
        categoryBreakdown:
          type: object
          additionalProperties:
            type: integer

    OnThisDay:
      type: object
      required: [oneWeekAgo, oneMonthAgo, oneYearAgo]
      properties:
        oneWeekAgo:
          oneOf:
            - $ref: "#/components/schemas/Entry"
            - type: "null"
        oneMonthAgo:
          oneOf:
            - $ref: "#/components/schemas/Entry"
            - type: "null"
        oneYearAgo:
          oneOf:
            - $ref: "#/components/schemas/Entry"
            - type: "null"

    Streak:
      type: object
      required: [currentStreak, longestStreak, todayLogged]
      properties:
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        todayLogged:
          type: boolean

    DailyPrompt:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string