# lib/api-zod/src/generated/types/authInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grateful – Gratitude Journal API
 * OpenAPI spec version: 0.2.0
 */

export interface AuthInput {
  email: string;
  /** @minLength 6 */
  password: string;
}