# lib/api-zod/src/generated/types/entry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grateful – Gratitude Journal API
 * OpenAPI spec version: 0.2.0
 */
import type { EntryMood } from './entryMood';

export interface Entry {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** 1-3 gratitude items */
  gratitudeItems: string[];
  reflection: string;
  /** @nullable */
  mood: EntryMood;
  starred: boolean;
  categories: string[];
  /** @nullable */
  photoPath?: string | null;
  createdAt: string;
}