# lib/api-zod/src/generated/types/entryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grateful – Gratitude Journal API
 * OpenAPI spec version: 0.2.0
 */
import type { EntryInputMood } from './entryInputMood';

export interface EntryInput {
  date: string;
  /**
     * @minItems 1
     * @maxItems 3
     */
  gratitudeItems: string[];
  reflection: string;
  /** @nullable */
  mood?: EntryInputMood;
  categories?: string[];
}