# lib/api-zod/src/generated/types/entryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grateful – Gratitude Journal API
 * OpenAPI spec version: 0.2.0
 */
import type { EntryUpdateMood } from './entryUpdateMood';

export interface EntryUpdate {
  /**
     * @minItems 1
     * @maxItems 3
     */
  gratitudeItems?: string[];
  reflection?: string;
  /** @nullable */
  mood?: EntryUpdateMood;
  starred?: boolean;
  categories?: string[];
}