# lib/api-zod/src/generated/types/onThisDay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grateful – Gratitude Journal API
 * OpenAPI spec version: 0.2.0
 */
import type { Entry } from './entry';

export interface OnThisDay {
  oneWeekAgo: Entry | null;
  oneMonthAgo: Entry | null;
  oneYearAgo: Entry | null;
}