# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grateful – Gratitude Journal API
 * OpenAPI spec version: 0.2.0
 */

export interface User {
  id: number;
  email: string;
  createdAt: string;
}