# lib/api-zod/src/generated/types/weeklySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grateful – Gratitude Journal API
 * OpenAPI spec version: 0.2.0
 */
import type { Entry } from './entry';
import type { WeeklySummaryCategoryBreakdown } from './weeklySummaryCategoryBreakdown';
import type { WeeklySummaryMoodBreakdown } from './weeklySummaryMoodBreakdown';

export interface WeeklySummary {
  startDate: string;
  endDate: string;
  entries: Entry[];
  totalEntries: number;
  moodBreakdown: WeeklySummaryMoodBreakdown;
  allGratitudeItems: string[];
  categoryBreakdown: WeeklySummaryCategoryBreakdown;
}